NoutBegin is not yielded as part of the "all*nouts" functions.
The reasoning is: the NoutBegin isn't really anything, it only serves as "nothing to see here"

The counterargument would be: we model NoutBegin as a referencable (via hash) Nout, so when
fetching all _nouts_ (as opposed to the notes), it's still "a thing".

Perhaps I'll revert this decision in the futute, we'll see.

--
This commit also renames 'follow_nouts' into 'all_preceding_nout_hashes'