PoC of multiple trees in a single window
I had to rewrite the color constants to be created on the fly.

The reason is that before the rewrite redraws in one of the histories would affect others.
In particular, other boxes would turn blue (which is the textcolor, not box color)

I can only surmise that this was caused by either OpenGL or Kivy throwing away the box_colors
on canvas.refresh()