PoC 'stable cursors for external changes'
When you have multiple windows, and something is changed in another window, this should not
affect what the cursor is pointing at. The exception is: if the other window deletes what you're
pointing at, you should fall back to the first still-exsiting ancestor.

This is implemented using t_addresses: receive_from_channel is implemented by keeping the
t_cursor the same (if possible) between before-and-after-update.

(Needs tests and some cleaning up still)