construct_y bugfix: don't improperly reuse 'structure'
The HistoriographyTreeNode was passed around, and reused, from the pre-replace to post-replace
construction. But there was no guarantee for a linear (or any) connection between those 2 things.

The current version doesn't pass around the value 'structure' anymore.

At the same time, the present solution makes it obvious that we can store all structures in a
big lookup table for caching (the nout_hash uniquely determines the HistoriographyTreeNode)