Make construct_x error-handling less agressive in "what constitutes an error"
Triggered by similar thoughts on construct_y.  The particular change is: it is now made explicit that failure on a
subtree does not propagate up.  A dissonent _does not_ break its parent: the parent's indexing scheme is still good, and
all actions in other subtrees are independent so they can continue.

To achieve this, the tree's error-ness was pushed into the nodes. This is asymetric with construct_y (where we keep
error-ness in a separate element in the tuple, and pass it around explicitly). I might change my mind back, we'll see.