Treenode brokenness: don't implement downward propagation of dissonents in the drawing mechanism
I.e. just because a parent is drawn in pink, doesn't mean all children should be drawn in pink.

If we were ever to change our minds to "downward propagation of dissonent-ness is
the correct decision", this decision should be implemented at the level of the
construction of the nodes, _not_ at the level of drawing them.

This commit also documents the current choices.