Lispy layout: always render child-0 as single-line
i.e. don't override user-override (which is currently broken).

Even though I imagine a version with user-override could be useful in the future, it's
currently broken.

Also: "always render child-0 as single-line" is our prime example of "rendering of the
parents affects rendering of the children"; I want such an example to exist before embarking
on a refactoring.