Remove s_address semantics from BoxNonTerminal
And reintroduce it as an annotation.

The idea being: the knowledge of where a node lives inside the total tree structure should not
influence the way the node is drawn; the exceptions to this rule should be formulated in the
form of InheritedRenderingInformation (future commit)

Cursor-display is temporarily broken as a consequence of this.