Refactor NT-drawing of trees into 2 phases (top down & bottom up pass)
The top down pass serves to pass rendering-information instructions from ancestors to descendants in the
tree. The prime example being: once you're inside a single-line-rendered node, all descendants are also
single-line.

The bottom-up pass constructs nodes based on the node itself, any inherited information and the already-rendered
children nodes. This is in contrast to the previous solution, which had the node-rendering call the children-rendering.

Because we now pass in the children-nodes, the way is paved to pass in exceptionally rendered nodes (future commit)

* is_cursor behavior was restored while doing the refactoring.