Viewports: store bounded information
Bounded means: no overscrolling and underscrolling of the document (scrolling pre-begin and past the end)

If you only bound to begin & end once you display, you can still "build up" invisible state. (as was the
case pre-this-commit). The funny behavior is then: if you hold CTRL+Y for a while, you must hold CTRL+E
for an equally long while before you see anything happening.

I actually thought that such "late bounding" could be a feature in the context of distributed editing. Example:
if you press 'zt' to put the cursor at the bottom of the screen, but this is not yet fully possible because you're
too close to the top of the document, and your document grows through another channel right after you pressed
'zt', the effects of this growth will be interpreted through the lens of 'zt'.

However, given the example above (and probably many others) I now stand corrected.