Show history widget's cursor again.
We'll need it soon, when we'll implement scrolling on the history widget.

It was turned off in 0ad6195c3474, when we removed the semantics of the s_path from the BoxNonTerminal
and reintroduced it as an annotation on those.

When doing that, we also removed the s_path from the recursive definition of draw_past_from_present. That
was strictly speaking not necessary.

Having a cursor visible is something I need now, so I just restored it.

At some point in the future we'll probably get rid of it again, with the purpose of being able to cache
larger textures (without any cursors on them);