ValueForm & BecomeValue take type information and a value of that type
The question whether to do this at analysis-time, or to simply pass the unparsed
string into the Note and Form was postponed until now.

The rationale for the present answer is simply: things that can be understood statically,
and/or even incrementally, should be understood at those time-frames.

The choice to model the type as an attribute of Value is somewhat arbitrary (an alternative
being: create a family of different Value types) but for now it's certainly good enough
of a choice.